ui: use inspire_categories for conference FoI.
ConferenceForm, `field_of_interest` field accepts inspire categories
unlike JobForm's arxiv_categories

INSPIR-2858
